implemented a global string with OnceCell
