+++ Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 tracing = "0.1.37"
 tracing-subscriber = {version = "0.3.16", features = ["fmt", "json", "env-filter", "registry"]}
+once_cell = "1.17.1"
 
 [dev-dependencies]
 cucumber = "0.19"
