+++ tests/minimal.rs
@@ -1,13 +1,9 @@
-use cucumber::{then, when, writer, World, WriterExt as _};
-use std::{
-    fs::{read_to_string, File},
-    io,
-    sync::Mutex,
-};
+use cucumber::{then, when, World};
 
-use test_of_trace;
 use tracing::{instrument, Level};
 use tracing_subscriber::fmt;
+
+use test_of_trace;
 
 #[derive(Debug, World)]
 #[world(init = Self::new)]
@@ -16,20 +12,18 @@ pub struct TestWorld {}
 impl TestWorld {
     #[instrument]
     fn new() -> Self {
-        let log_file = File::create("my_temp_trace.log").unwrap();
         fmt()
             .without_time()
             .with_line_number(true)
             .with_max_level(Level::TRACE)
             .json()
-            .with_writer(Mutex::new(log_file))
+            .with_writer(test_of_trace::logger::Logger::new)
             .init();
         Self {}
     }
     fn level_contains(&self, level: &Level, pattern: &str) -> Result<bool, std::io::Error> {
-        let buffer = read_to_string("my_temp_trace.log")?;
-        Ok(buffer
-            .lines()
+        let s = test_of_trace::logger::GLOBAL_STRING.lock().unwrap();
+        Ok(s.lines()
             .filter(|l| l.starts_with(&("{\"level\":\"".to_owned() + level.as_str())))
             .any(|l| l.contains(pattern)))
     }
@@ -42,7 +36,6 @@ fn call_add(_world: &mut TestWorld) {
 
 #[then("a trace is emitted")]
 fn trace_is_emitted(world: &mut TestWorld) {
-    //TODO
     match world.level_contains(&Level::INFO, "result is") {
         Ok(val) => assert!(val),
         Err(_) => assert!(false),
@@ -52,14 +45,4 @@ fn trace_is_emitted(world: &mut TestWorld) {
 #[tokio::main]
 async fn main() {
     TestWorld::run("tests/features").await;
-    // for debugging output:
-    //TestWorld::cucumber()
-    //.max_concurrent_scenarios(1)
-    //.with_writer(
-    //writer::Basic::raw(io::stdout(), writer::Coloring::Never, 0)
-    //.summarized()
-    //.assert_normalized(),
-    //)
-    //.run("tests/features")
-    //.await;
 }
